pyproject.toml
[project]
name = "ecommerce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cloudinary>=1.44.1",
    "django>=6.0.2",
    "django-cloudinary-storage>=0.3.0",
    "django-jazzmin>=3.0.3",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "mysqlclient>=2.2.8",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
